fnr-analytics/requirements.txt
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pyarrow>=15.0.0
